Refactor attr/effect code into base class
